In-place factories moved from /detail to /utility New value_initalized tests added
git-svn-id: http://svn.boost.org/svn/boost/trunk@25436 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/libs/optional/test/optional_test_ref_fail1.cpp
@@ -17,8 +17,8 @@
 void optional_reference__test_no_converting_assignment()
 {
   boost::optional<int&> opt ;
-  short v = 1 ;
-  short& r = v ;
+  double v = 1 ;
+  double& r = v ;
   opt = r ;
 }
